+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @blocksuite/blocks
 
+## 0.2.22
+
+### Patch Changes
+
+- 9927578: fix hotkey on windows
+- Updated dependencies [9927578]
+  - @blocksuite/store@0.2.22
+
 ## 0.2.19
 
 ### Patch Changes